Arregle un error en los experimentos.

+++ codigo/experimentacion/exp1.c
@@ -200,7 +200,7 @@ void correr_exp(enum experimentos exp, int tm, int step, int n, int ldr_alpha, i
                 }
                 mediciones[k] -= start;
             }
-            qsort(mediciones,n, sizeof(int),compare_function);
+            qsort(mediciones,n, sizeof(long long int),compare_function);
             for( k = 0; k < na; k++) {
                 avg += mediciones[k];
             }
@@ -210,8 +210,8 @@ void correr_exp(enum experimentos exp, int tm, int step, int n, int ldr_alpha, i
             }
             sd = sd / na;
             sd = sqrt(sd);
-            fprintf(fa, "%llu, ", avg / (i*j));
-            fprintf(fs, "%llu, ", sd / (i*j));
+            fprintf(fa, "%f, ", ((double) avg) / ((double) (i*j)));
+            fprintf(fs, "%f, ", ((double) sd) / ((double) (i*j)));
             bmp_delete(img);
             bmp_delete(imgD);
         }
